requirements.txt

mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.26.4
pandas==2.0.3
pandas-gbq==0.23.1
requests-oauthlib==2.0.0
streamlit==1.37.0








